merge in

+++ Slim/Plugin/OnlineLibrary/Importer.pm
@@ -1,6 +1,6 @@
 package Slim::Plugin::OnlineLibrary::Importer;
 
-# Logitech Media Server Copyright 2001-2020 Logitech.
+# Logitech Media Server Copyright 2001-2023 Logitech.
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License,
 # version 2.
@@ -29,7 +29,7 @@ sub initPlugin {
 		type   => 'post',
 		weight => 110,       # this importer needs to be run after the VirtualLibraries (weight: 100)
 		onlineLibraryOnly => 1,
-		'use'  => 1,
+		'use'  => $prefs->get('enablePreferLocalLibraryOnly'),
 	} );
 
 	my $mappings = $prefs->get('genreMappings');
@@ -154,27 +154,68 @@ package Slim::Plugin::OnlineLibrary::Importer::VirtualLibrariesCleanup;
 sub startScan {
 	my ($class) = @_;
 
+	if (!$prefs->get('enablePreferLocalLibraryOnly')) {
+		Slim::Music::Import->endImporter($class);
+		return;
+	}
+
 	my $dbh = Slim::Schema->dbh or return;
 
 	$dbh->do('DROP TABLE IF EXISTS duplicate_albums');
 	$dbh->do(q(
 		CREATE TEMPORARY TABLE duplicate_albums AS
-			SELECT albums.id
+			SELECT albums.id AS online, otheralbums.id AS local
 			FROM albums
-			WHERE albums.extid IS NOT NULL
-				AND 1 IN (
-					SELECT 1
-					FROM albums otheralbums
-					WHERE otheralbums.extid IS NULL
-						AND LOWER(otheralbums.title) = LOWER(albums.title)
-						AND otheralbums.contributor = albums.contributor
-				)
+			JOIN albums otheralbums ON otheralbums.extid IS NULL AND
+				LOWER(otheralbums.title) = LOWER(albums.title) AND
+				otheralbums.contributor = albums.contributor
+			WHERE albums.extid IS NOT NULL AND
+				(
+					SELECT otheralbums.id
+						FROM albums otheralbums
+						WHERE otheralbums.extid IS NULL AND
+								LOWER(otheralbums.title) = LOWER(albums.title) AND
+								otheralbums.contributor = albums.contributor
+				) IS NOT NULL
+	) );
+
+	$dbh->do('CREATE INDEX IF NOT EXISTS online ON duplicate_albums (online)');
+	$dbh->do('CREATE INDEX IF NOT EXISTS local ON duplicate_albums (local)');
+
+	my $delAlbumSth = $dbh->prepare_cached(q(
+		DELETE FROM library_album
+		WHERE library_album.album = ? AND library = ? AND (
+			SELECT online FROM duplicate_albums WHERE online = ?
+		) IS NOT NULL
 	));
 
-	$dbh->do(q(DELETE FROM library_album WHERE library_album.album IN (SELECT id FROM duplicate_albums)));
-	$dbh->do(q(DELETE FROM library_track WHERE library_track.track in (
-		SELECT id FROM tracks WHERE tracks.album IN (SELECT id FROM duplicate_albums)
-	)));
+	my $onlineAlbumsSth = $dbh->prepare_cached(q(
+		SELECT album FROM library_album WHERE library = ? AND album IN (SELECT online FROM duplicate_albums)
+	));
+
+	my $localAlbumsSth = $dbh->prepare_cached(q(
+		SELECT album FROM library_album WHERE library = ? AND album IN (SELECT local FROM duplicate_albums WHERE online = ?)
+	));
+
+	my $libraryAlbumDelSth = $dbh->prepare_cached(q(
+		DELETE FROM library_album WHERE library = ? AND album = ?
+	));
+
+	my $libraryTrackDelSth = $dbh->prepare_cached(q(
+		DELETE FROM library_track WHERE library = ? AND track IN (SELECT id FROM tracks WHERE album = ?)
+	));
+
+	foreach my $library (keys %{ Slim::Music::VirtualLibraries->getLibraries() } ) {
+		$onlineAlbumsSth->execute($library);
+		while ( my ($onlineId) = $onlineAlbumsSth->fetchrow_array ) {
+			$localAlbumsSth->execute($library, $onlineId);
+			my $localAlbumIds = $localAlbumsSth->fetchall_arrayref;
+			foreach my $albumId (@{ $localAlbumIds || [] }) {
+				$libraryAlbumDelSth->execute($library, $albumId->[0]);
+				$libraryTrackDelSth->execute($library, $albumId->[0]);
+			}
+		}
+	}
 
 	$dbh->do('DROP TABLE IF EXISTS duplicate_albums');
